network/dnstruncate: split packet relay implementation and tests from packet proxy
